Add files via upload
Title: COVID Exploratory Analysis: SQL Portfolio Project:

Description:
This SQL portfolio project provides an in-depth exploratory analysis of COVID-19 data, showcasing the power of SQL for data analysis and visualization. The project is available on GitHub and serves as a valuable resource for anyone interested in understanding and analyzing the impact of the COVID-19 pandemic.

The project utilizes a comprehensive dataset that includes worldwide COVID-19 statistics, encompassing infection rates, deaths, recoveries, and other relevant metrics. By leveraging SQL queries and techniques, the project aims to uncover meaningful insights and trends hidden within the data, allowing users to gain a deeper understanding of the global pandemic.